% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{allocate}
\alias{allocate}
\title{ssuex: SSUs in Exeter}
\usage{
allocate(sFile, uFile, oFile)
}
\arguments{
\item{sFile}{Tab delimited file with header containing information about students:
\code{student name, number, c1, ..., cN}}

\item{uFile}{Tab delimited file with header containing information about SSUs:
\code{unit code, spaces available}}

\item{oFile}{Name of output file}
}
\value{
NA
}
\description{
This function allocates SSUs to students based on their preferences.
It takes two tab-delimited headed files containing student preferences and
information about available SSUs and returns an output file containing the
the assigned SSU.
}
